refactored feature testing view
